feat(auth+admin): complete responsive admin login and setup protected admin dashboard layout
