+++ package.json
@@ -15,6 +15,7 @@
     "node-fetch": "^3.3.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-relay": "^19.0.0",
     "relay-runtime": "^19.0.0"
   },
