+++ src/app/(admin)/dashboard/page.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import React from 'react';
+
+const DashboardPage: React.FC = () => {
+
+    return (
+        <div className="w-full h-full flex items-center justify-center px-12">
+            In dashboard Page
+        </div>
+    );
+};
+
+export default DashboardPage;
